+++ backend/server.py
@@ -464,6 +464,7 @@ from routes.kodefikasi import kodefikasi_router
 from routes.persediaan import persediaan_router
 from routes.persediaan_laporan import persediaan_laporan_router
 from routes.persediaan_permohonan import persediaan_permohonan_router
+from routes.aset_permohonan import aset_permohonan_router
 from routes.penggunaan import penggunaan_router
 from routes.pengamanan import pengamanan_router
 from routes.pemeliharaan import pemeliharaan_router
@@ -527,6 +528,7 @@ api_router.include_router(kodefikasi_router)
 api_router.include_router(persediaan_router)
 api_router.include_router(persediaan_laporan_router)
 api_router.include_router(persediaan_permohonan_router)
+api_router.include_router(aset_permohonan_router)
 api_router.include_router(penggunaan_router)
 api_router.include_router(pengamanan_router)
 api_router.include_router(pemeliharaan_router)